+++ graal-js/src/com.oracle.truffle.js.runtime/src/com/oracle/truffle/js/runtime/builtins/JSGlobalObject.java
@@ -52,7 +52,8 @@ public final class JSGlobalObject extends JSBuiltinObject {
 
     public static final String CLASS_NAME = "global";
     public static final String EVAL_NAME = "eval";
-    public static final String CLASS_NAME_NASHORN_EXTENSIONS = "globalNashornExtensions";
+    public static final String CLASS_NAME_NASHORN_EXTENSIONS = "globalNashornScripting";
+    public static final String CLASS_NAME_SHELL_EXTENSIONS = "globalShell";
 
     public static final JSGlobalObject INSTANCE = new JSGlobalObject();