+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The main focus is on user-observable behavior of the engine.
 * `Atomics.wake` available under its new name (`Atomics.notify`) as well.
 * [Well-formed JSON.stringify](https://github.com/tc39/proposal-well-formed-stringify) proposal implemented.
 * [globalThis](https://github.com/tc39/proposal-global) proposal implemented.
+* Disabled non-standard global functions `quit`, `read`, `readbuffer`, and `readline` by default, except for the `js` launcher (`js.shell` option).
+* Disabled non-standard global functions `readFully` and `readLine` by default, now available only in Nashorn scripting mode (`--js.scripting`).
+* Disabled Nashorn syntax extensions by default (`js.syntax-extensions` option), barring `ScriptEngine`.
+* Note: As a result, `eval` requires function expressions to be parenthesized, e.g.: `eval("(function(){...})")` (not `eval("function(){...}")`).
 
 ## Version 1.0.0 RC8
 * Provide simplified implementations for methods of the global `console` object even outside Node.js mode.